PropertEz/PropertEz-Backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
Django==5.0
django-cors-headers==4.3.1
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
